// src/app/models/client.model.ts
export interface Client {
        id: number;
        name: string;
        email: string;
        city: string;
        address: string;
        irrigated: number;
        noirrigated: number;
        total: number;
        homogenous: number;
        heterogenous: number;
        clay: number;
        silt: number;
        sand: number;
        limestone: number;
        gypsum: number;
        maps: string;
      
}